Create my_brackets.c
|Brackets||

Description
Write a function that one string as argument. The function returns 1
if the expression is correctly bracketed, otherwise it returns 0.

Symbols considered as 'brackets' are brackets '(' and ')', square brackets '['
and ']'and braces '{' and '}'. Every other symbols are simply ignored.

An opening bracket must always be closed by the good closing bracket in the
correct order. A string which not contains any bracket is considered as a
correctly bracketed string.

Example 00:

Input: "(johndoe)"
Output: 1
Example 01:

Input: "([)]"
Output: 0
Example 02:

Input "{[(0 + 0)(1 + 1)](3*(-1)){()}}"
Output: 1
Example 03:

Input ""
Output: 1