Create extand_string.c
|Expand String||

Description
Write a program that takes a string and displays it with exactly three spaces
between each word, with no spaces or tabs either at the beginning or the end,
followed by a newline.

A word is a section of string delimited either by spaces/tabs, or by the
start/end of the string.

Example 00:

Input: "See? It's easy to print the same thing"
Output: "See?   It's   easy   to   print   the   same   thing"
Example 01:

Input: " this        time it      will     be    more complex  . "
Output: "this   time   it   will   be   more   complex"
Example 02:

Input "No  S***    Sherlock..."
Output: "No   S***   Sherlock..."
Example 03:

Input ""
Output: ""