Create my_strstr
|My Strstr||

Description
Implement strStr().

Returns a pointer to the first occurrence of needle in haystack, or NULL if needle is not part of haystack.

Clarification:

What should we return when needle is an empty string? This is a great question to ask during an interview.

For the purpose of this problem, we will return haystack when needle is an empty string.